* Refactoring * Cleaning * Progress on physics dataset
