+++ tools/ml/settings/2021-08-23T16:35:40.893219.conf
@@ -25,7 +25,6 @@ TENSORBOARD_DEBUGGING=False
 LATENT_DIMENSION=2
 
 [Data]
-#DATA1=generator-ImageDataset
 DATA1=generator-MnistDataset
 
 [Plotting]
