+++ tools/ml/src/custom_datasets/s3_image_dataset.py
